refactor: simplify cow amm banner
