+++ apps/cowswap-frontend/src/common/pure/ArrowBackground/index.tsx
