+++ apps/cowswap-frontend/src/common/pure/CoWAmmBannerContent/dummyData.ts
