+++ apps/cowswap-frontend/src/common/pure/CoWAmmBannerContent/types.ts
@@ -1,0 +1,4 @@
+export enum BannerLocation {
+  Global = 'global',
+  TokenSelector = 'tokenSelector',
+}
